refactor(import): reduce bundle size by import hooks one-by-one
