+++ src/hooks/useClipboard.ts
@@ -1,0 +1,7 @@
+import useCopyToClipboard from 'react-use/lib/useCopyToClipboard'
+
+export type { CopyToClipboardState as TCopyToClipboardState } from 'react-use/lib/useCopyToClipboard'
+
+// export type TCopyToClipboardState = CopyToClipboardState
+
+export default useCopyToClipboard
