+++ src/hooks/useInterval.ts
@@ -1,0 +1,3 @@
+import useInterval from 'react-use/lib/useInterval'
+
+export default useInterval
