+++ src/hooks/useSwipe.ts
@@ -1,0 +1,5 @@
+import { useSwipeable } from 'react-swipeable'
+
+const useSwipe = useSwipeable
+
+export default useSwipe
